rosa/requirements.txt
boto
dateparser